Revert LockedMenuItem unification - keep separate for pages and snippets
